requirements.txt
streamlit
numpy
pandas
scikit-learn
umap-learn
plotly
